refactor(solver): split function generic constraint helpers
